backend/requirements.txt
fastapi
uvicorn
python-dotenv
sentence-transformers
faiss-cpu
groq
requests
PyPDF2
python-multipart
